improve stock watch

+++ BlazorMarketWatch/BlazorMarketWatch.Web/Components/Pages/YourWatchBase.cs
@@ -1,7 +1,10 @@
 ﻿using BlazorMarketWatch.Web.Entities;
 using BlazorMarketWatch.Web.Repositories.Contracts;
+using BlazorMarketWatch.Web.Services;
+using BlazorMarketWatch.Web.Services.Contracts;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Authorization;
+using static BlazorMarketWatch.Web.Dtos.StockDto;
 
 namespace BlazorMarketWatch.Web.Components.Pages
 {
@@ -13,11 +16,19 @@ namespace BlazorMarketWatch.Web.Components.Pages
         [Inject]
         public AuthenticationStateProvider authenticationStateProvider { get; set; }
 
+        [Inject]
+        public IStockService StockService { get; set; }
+
+        public TickerSummaryDto TickerSummaryData { get; set; }
+
         public IEnumerable<UserStock> UserStocks { get; set; }
+
+        public IEnumerable<TickerSummary> Summaries { get; set; }
         public string ErrorMessage { get; set; }
 
         protected override async Task OnInitializedAsync()
         {
+            TickerSummaryData = await StockService.GetStockTickers();
             GetUserStocks();
         }
 
@@ -31,12 +42,31 @@ namespace BlazorMarketWatch.Web.Components.Pages
                 {
                     throw new Exception("Error fetching user stocks");
                 }
+                Summaries = ConvertToTickerSummary(UserStocks);
+                if (Summaries == null)
+                {
+                    throw new Exception("Error converting tickers to summaries");
+                }
                 StateHasChanged();
             }
             catch (Exception ex)
             {
                 ErrorMessage = ex.Message;
             }
+        }
+
+        private IEnumerable<TickerSummary> ConvertToTickerSummary(IEnumerable<UserStock> tickers)
+        {
+            List<TickerSummary> result = new List<TickerSummary>();
+            foreach (UserStock ticker in tickers)
+            {
+                var summary = TickerSummaryData.data.Where(x => x.symbol == ticker.Ticker).FirstOrDefault();
+                if (summary != null)
+                {
+                    result.Add(summary);
+                }
+            }
+            return result;
         }
 
         async Task<string> getUserId()
@@ -46,10 +76,15 @@ namespace BlazorMarketWatch.Web.Components.Pages
             return UserId;
         }
 
-        protected async void RemoveStock_Click(int id)
+        protected async void RemoveStock_Click(string ticker)
         {
             try
             {
+                var id = ConvertTickerToId(ticker);
+                if (id == 0)
+                {
+                    throw new Exception("Remove conversion failed");
+                }
                 var result = await userStockRepository.DeleteUserStock(id);
                 if (result == null)
                 {
@@ -61,6 +96,16 @@ namespace BlazorMarketWatch.Web.Components.Pages
             {
                 ErrorMessage = ex.Message;
             }
+        }
+
+        private int ConvertTickerToId(string ticker)
+        {
+            var stock = UserStocks.Where(x => x.Ticker == ticker).FirstOrDefault();
+            if (stock != null)
+            {
+                return stock.Id;
+            }
+            return 0;
         }
     }
 }
